requirements.txt
numpy
pandas
yfinance
python-dotenv
requests
azure-functions
azure-storage-blob
python-telegram-bot